+++ src/test/groovy/ru/vyarus/guice/persist/orient/db/scheme/initializer/ext/field/index/fulltext/FulltextIndexModel.groovy
@@ -9,9 +9,6 @@ class FulltextIndexModel {
     @FulltextIndex
     String defaults
 
-    @FulltextIndex(useHashIndex = true)
-    String hash
-
     @FulltextIndex(name = "all_options",
             indexRadix = false,
             ignoreChars = "'",