+++ CHANGELOG.md
@@ -1,5 +1,5 @@
 * Update to guice 4.2.2
-* (breaking) Update to Orient 3.0.11
+* (breaking) Update to Orient 3.0.38
     - Bound connection objects changed:
         - Provider<OObjectDatabaseTx> -> Provider<ODatabaseObject>
         - Provider<ODatabaseDocumentTx> -> Provider<ODatabaseDocument>
@@ -22,7 +22,10 @@
         - Removed pool recovery logic (since 1.x: when pool failed to provide correct connection it was re-created).
             Assuming new pool is more stable.             
         - Different user credentials (UserManager.executeWithUser) will create separate connection outside the pool.
-    - Object api now may be used to properly remove graph nodes with edges (because graph consistency is on document level now)    
+    - Object api now may be used to properly remove graph nodes with edges (because graph consistency is on document level now)
+    - Remove useHashIndex attribute in @FullTextIndex schema annotation 
+      (OClass.INDEX_TYPE.FULLTEXT_HASH_INDEX removed in orient 3.0.38:
+      https://github.com/orientechnologies/orientdb/commit/bfceffa50d3f708f5c1c05dab1f082861df01e12#diff-3371617e7407306ad4397a0835f64175314b828295d5ad88891c051915d2d8aaL226)
 
 ### 3.3.2 (2018-04-02)
 * Guice 4.2.0 compatibility